NOW zPos takes in worldSize :D

+++ Item.swift
@@ -47,7 +47,7 @@ public class Item : SKSpriteNode {
 	//----------------------------------------------------------------
 	
 	func updateZPosition () {
-		zPosition = (worldSize * 4) - (cartesianPoint.x + cartesianPoint.y)
+		zPosition = (CGFloat(worldSize) * 4) - (cartesianPoint.x + cartesianPoint.y)
 	}
 	
 }
